services/api_gateway/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api_gateway"
version = "0.1.0"
description = "API Gateway service - API Gateway microservice"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.123.5",
    "grpcio>=1.76.0",
    "pydantic>=2.12.5",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
api-gateway = "api_gateway.cli:run"
[tool.setuptools.packages.find]
where = ["."]
include = ["api_gateway", "api_gateway.*"]

[tool.uv.sources]
shared = { workspace = true }
